backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
pillow
python-dotenv
requests
mistralai
ipykernel
openai
langgraph
langchain